--- CodepenChallenge Teach JS/Project React/src/css/AppButtons.css
@@ -1,17 +1,0 @@
-/* create a grid with one column and implicit rows matching the column size */
-.AppButtons {
-  display: grid;
-  grid-gap: 12px;
-  grid-template-columns: 1fr;
-  grid-auto-rows: auto;
-  /*  
-  include auto overflow for the grid responsible for the buttons
-  effectively this squishes the width, which requires additional space to visually match the height 
-  */
-  overflow-y: auto;
-}
-.AppButtons svg {
-  width: 100%;
-  height: 100%;
-  pointer-events: none;
-}
