+++ CodepenChallenge Teach JS/Project React/src/css/GridButtons.css
@@ -1,0 +1,21 @@
+/* display the buttons in a single column */
+.AppButtons {
+  display: grid;
+  grid-gap: 15px;
+  grid-template-columns: 1fr;
+  /* grid-auto-rows: auto; */
+  /*  
+  include auto overflow for the grid responsible for the buttons
+  effectively this squishes the width, which requires additional space to visually match the height 
+  */
+  overflow-y: auto;
+}
+/*
+stretch the SVG to cover the buttons' width and height 
+remove pointer events to allow click events on the buttons instead
+*/
+.AppButtons svg {
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+}
